Improve Slack notification logic for better filtering
