+++ helpers/notifications.ts
@@ -30,26 +30,13 @@ export interface AgentNotificationOptions {
   customLinks?: string;
 }
 
-function shouldSkipNotification(options: SlackNotificationOptions): boolean {
-  const jobStatus = options.jobStatus || "failed";
-
-  if (jobStatus === "success") {
-    console.log(`Slack notification skipped (status: ${jobStatus})`);
-    return true;
-  }
-
-  const branchName = (process.env.GITHUB_REF || "").replace("refs/heads/", "");
-  if (branchName !== "main" && process.env.GITHUB_ACTIONS) {
-    console.log(`Slack notification skipped (branch: ${branchName})`);
-    return true;
-  }
-
-  return false;
-}
-
-function generateMessage(options: SlackNotificationOptions): string {
+export async function sendSlackNotification(
+  options: SlackNotificationOptions,
+  channel?: string,
+): Promise<void> {
+  const targetChannel = channel || process.env.SLACK_CHANNEL || "general";
   const testName = options.testName
-    ? options.testName[0].toUpperCase() + options.testName.slice(1)
+    ? options.testName[0].toUpperCase() + options.testName.slice(1) + " - "
     : "";
 
   const errorLogsArr = Array.from(options.errorLogs || []);
@@ -64,12 +51,7 @@ function generateMessage(options: SlackNotificationOptions): string {
     `Logs:\n\`\`\`${logs}\`\`\``,
   ];
 
-  return sections.filter(Boolean).join("\n");
-}
-
-async function postToSlack(message: string, channel?: string): Promise<void> {
-  const targetChannel = channel || process.env.SLACK_CHANNEL || "general";
-
+  const message = sections.filter(Boolean).join("\n");
   const response = await fetch("https://slack.com/api/chat.postMessage", {
     method: "POST",
     headers: {
@@ -92,50 +74,5 @@ async function postToSlack(message: string, channel?: string): Promise<void> {
     console.log(`✅ Slack notification sent successfully to ${targetChannel}!`);
   } else {
     console.error("❌ Failed to send Slack notification. Response:", data);
-  }
-}
-
-export async function sendSlackNotification(
-  options: SlackNotificationOptions,
-): Promise<void> {
-  if (!process.env.SLACK_BOT_TOKEN) {
-    console.log("Slack notification skipped (SLACK_BOT_TOKEN not set)");
-    return;
-  }
-  if (options.label === "error") {
-    if (!options.errorLogs || options.errorLogs.size === 0) {
-      console.log(
-        "Slack notification skipped (no actual test failures detected)",
-      );
-      return;
-    }
-
-    if (shouldSkipNotification(options)) {
-      return;
-    }
-  }
-
-  if (options.errorLogs && options.errorLogs.size > 0) {
-    const failLines = extractFailLines(options.errorLogs);
-    await sendDatadogLog(Array.from(options.errorLogs), {
-      channel: options.channel,
-      test: options.testName,
-      failLines: Array.from(failLines).length,
-      env: process.env.ENVIRONMENT || process.env.XMTP_ENV,
-      region: process.env.GEOLOCATION,
-      sdk: "latest",
-    });
-  }
-
-  // Check if test should be filtered out
-  if (options.errorLogs && shouldFilterOutTest(options.errorLogs)) {
-    return;
-  }
-
-  try {
-    const message = generateMessage(options);
-    await postToSlack(message, options.channel);
-  } catch (error) {
-    console.error("Error sending Slack notification:", error);
   }
 }
